Move job-handling changesets to a submodule.
