+++ deployment/data-streams/changeset/jobs/common.go
@@ -1,4 +1,4 @@
-package changeset
+package jobs
 
 import (
 	"context"
