+++ deployment/data-streams/changeset/jobs/jd_distribute_llo_jobs.go
@@ -1,4 +1,4 @@
-package changeset
+package jobs
 
 import (
 	"context"
