+++ deployment/data-streams/changeset/jobs/jd_distribute_stream_jobs_test.go
@@ -1,4 +1,4 @@
-package changeset
+package jobs
 
 import (
 	"testing"
