+++ deployment/data-streams/changeset/jobs/jd_revoke_jobs_test.go
@@ -1,4 +1,4 @@
-package changeset
+package jobs
 
 import (
 	"regexp"
